requirements.txt
python-telegram-bot==20.3
telethon==1.34.0
nest_asyncio
